Develop

+++ include/index/ranker/absolute_discount.h
@@ -20,6 +20,17 @@ namespace index
 
 /**
  * Implements the absolute discounting smoothing method.
+ *
+ * Required config parameters:
+ * ~~~toml
+ * [ranker]
+ * method = "absolute-discount"
+ * ~~~
+ *
+ * Optional config parameters:
+ * ~~~toml
+ * delta = 0.7
+ * ~~~
  */
 class absolute_discount : public language_model_ranker
 {
@@ -32,23 +43,23 @@ class absolute_discount : public language_model_ranker
     /**
      * @param delta
      */
-    absolute_discount(double delta = 0.7);
+    absolute_discount(float delta = 0.7f);
 
     /**
      * Calculates the smoothed probability of a term.
      * @param sd score_data for the current query
      */
-    double smoothed_prob(const score_data& sd) const override;
+    float smoothed_prob(const score_data& sd) const override;
 
     /**
      * A document-dependent constant.
      * @param sd score_data for the current query
      */
-    double doc_constant(const score_data& sd) const override;
+    float doc_constant(const score_data& sd) const override;
 
   private:
     /// the absolute discounting parameter
-    const double delta_;
+    const float delta_;
 };
 
 /**
